Worked on (and finished) database processing and information retrieval. Added Ajax to the front-end, and did a lot of work on overall design.

+++ models/user.js
@@ -29,7 +29,7 @@ module.exports = sequelize.define('user', {
     type: Sequelize.INTEGER
   },
   days_spent_watching: {
-    type: Sequelize.INTEGER
+    type: Sequelize.DECIMAL(5,2)
   }
 }, {
   timestamps: true,
